feature 29: отправка сообщений по email
