+++ src/bot/handlers/command_application.py
@@ -9,10 +9,11 @@ from telegram.ext import (
     filters,
 )
 
+from bot.constants.schemas import QuestionModel
 from bot.constants.state import GET_USER_QUESTION
 from bot.constants.text import BACK_TO_MENU, STOP_MESSAGE
 from bot.keyboards.keyboards import main_menu_markup
-from bot.utils import QuestionModel, send_email
+from bot.utils import send_email
 
 
 async def help_callback(
